requirements.txt
flask
flask-login
flask-peewee
flask-wtf
flask-mail
wtforms
flask-script
uwsgi
markdown
